% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTools.R
\name{graphTools}
\alias{graphTools}
\title{Functions to plot DESeq2 results}
\usage{
graphTools(
  res,
  maxpadj,
  minlfc,
  device = "png",
  output_folder = ".",
  width = 7,
  height = 7,
  plot.title = NULL
)
}
\arguments{
\item{res}{Data frame object with DESeq2 results}

\item{maxpadj}{P-adjusted value for significant features}

\item{minlfc}{Value for dysregulated features}

\item{device}{Format for graphs ("pdf", "svg", "eps", "png", "tiff", "jpeg").
A vector for several formats can be uses: c("svg", "png")}

\item{output_folder}{Folder where graphs will be saved}

\item{width}{Graph width in inches}

\item{height}{Graph height in inches}

\item{plot.title}{Text for graph titles}
}
\description{
Functions to plot DESeq2 results
}
\examples{
\dontrun{
  graphTools(res, maxpadj, minlfc, device, 
             output, plot.title = plot.title)
}

}
