% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTools.R
\name{printDevice}
\alias{printDevice}
\title{Save ggplot to Multiple File Formats}
\usage{
printDevice(plot, basename, device, width = 7, height = 7)
}
\arguments{
\item{plot}{A ggplot object to save.}

\item{basename}{Character string. Base name for output files (without extension).
Extensions will be added automatically based on requested formats.}

\item{device}{Character vector. File format(s) to save. Supported formats:
"svg", "eps", "png", "tiff", "jpeg". Case-insensitive. Dots are
automatically removed (e.g., ".png" becomes "png").}

\item{width}{Numeric. Plot width in inches. Default is 7.}

\item{height}{Numeric. Plot height in inches. Default is 7.}
}
\description{
Internal function to save a ggplot object to one or more file formats.
Supports common vector (SVG, EPS) and raster (PNG, TIFF, JPEG) formats.
}
