% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTEcounts.R
\name{readTEcounts}
\alias{readTEcounts}
\title{Read TE count data and merge in a data frame by sample}
\usage{
readTEcounts(datafile, folder)
}
\arguments{
\item{datafile}{A tab separated file with four columns and a header. Structure
must follow this example:

File         Sample    Group     Condition
WT1.cntTable    WT1        WT       Control
....
KO1.cntTable    KO1        KO       Treat

"Control" and "Treat" labels are required for DESeq2 analysis.}

\item{folder}{Folder for count files listed in first column}
}
\value{
Data frame with counts by features in rows and samples in columns
}
\description{
Read TE count data and merge in a data frame by sample
}
\examples{
datafile <- "my.path/data.csv"
folder <- "data/"
df <- readTEexpress(datafile, folder)
}
