% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_DEA.R
\name{TE_DEA}
\alias{TE_DEA}
\title{Compute differential expression analysis for transposable elements}
\usage{
TE_DEA(
  metafile,
  folder,
  output = ".",
  useCtrlGenes = FALSE,
  shrinklog2FC = FALSE
)
}
\arguments{
\item{metafile}{Full path of metadata file}

\item{folder}{Full path for folder of count files}

\item{output}{Path for output folder}

\item{useCtrlGenes}{Boolean for using only genes in estimating DESeq2 size factors (default: FALSE)}

\item{shrinklog2FC}{Boolean for applying log2FC shrinking in DESEq2 (default: FALSE)}
}
\value{
Table of results
}
\description{
Compute differential expression analysis for transposable elements
}
